move file out of scipts
