+++ catalog_reader/version_bump.py
